requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic-settings==2.6.0
pytest==8.3.3
croniter==3.0.3

